requirements.txt
absl-py==0.11.0
altair==4.1.0
appnope==0.1.0
argon2-cffi==20.1.0
astor==0.8.1
astroid==2.4.2
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
base58==2.1.0
beautifulsoup4==4.9.3
bert-for-tf2==0.14.9
bleach==3.2.1
blinker==1.4
branca==0.4.2
Brotli==1.0.9
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
chart-studio==1.1.0
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.4
colorlover==0.3.0
confuse==1.4.0
cufflinks==0.17.3
cycler==0.10.0
Cython==0.29.14
dash==1.19.0
dash-core-components==1.15.0
dash-html-components==1.1.2
dash-renderer==1.9.0
dash-table==4.11.2
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
et-xmlfile==1.0.1
Flask==1.1.2
Flask-Compress==1.8.0
Flask-HTTPAuth==4.2.0
Flask-SQLAlchemy==2.4.4
flatbuffers==1.12
folium==0.12.1
future==0.18.2
gast==0.3.3
gensim==3.8.3
gitdb==4.0.5
GitPython==3.1.14
google-auth==1.27.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
html5lib==1.1
htmlmin==0.1.12
httpauth==0.3
idna==2.10
ImageHash==4.2.0
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.6.4
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.17.2
Jinja2==2.11.2
joblib==1.0.0
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
llvmlite==0.35.0
lxml==4.6.2
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
mccabe==0.6.1
missingno==0.4.2
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
networkx==2.5
nltk==3.5
notebook==6.2.0
numba==0.52.0
numpy==1.19.3
oauthlib==3.1.0
opencv-python==4.5.1.48
openpyxl==3.0.5
opt-einsum==3.3.0
packaging==20.8
pandas==1.1.4
pandocfilters==1.4.3
params-flow==0.8.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
plotly==4.14.1
prometheus-client==0.9.0
prompt-toolkit==3.0.8
protobuf==3.15.1
psutil==5.8.0
ptyprocess==0.6.0
py-params==0.10.2
pyarrow==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydeck==0.6.1
Pygments==2.7.2
pylint==2.6.0
PyMySQL==1.0.2
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.4
PyWavelets==1.1.1
pywin32==300
pywinpty==0.5.7
pyzmq==20.0.0
regex==2020.11.13
requests==2.25.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7.1
scikit-learn==0.24.1
scipy==1.5.4
seaborn==0.11.0
Send2Trash==1.5.0
sentencepiece==0.1.95
shap==0.38.1
six==1.15.0
sklearn==0.0
slicer==0.0.7
smart-open==4.2.0
smmap==3.0.5
sodapy==2.1.0
soupsieve==2.1
tangled-up-in-unicode==0.0.6
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.2
toolz==0.11.1
tornado==6.1
tqdm==4.56.0
traitlets==5.0.5
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.2
validators==0.18.2
visions==0.6.0
watchdog==2.0.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wordcloud==1.8.1
